microservices/ocr/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0
python-dotenv==1.0.1
